backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
slowapi==0.1.9
pydantic==2.9.2
pydantic-settings==2.6.0
MetaTrader5==5.0.4778
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
scipy==1.14.1
python-dotenv==1.0.1
httpx==0.27.2
websockets==13.1
loguru==0.7.2
